Added C++ code for reading the Campbell Ceilometer
This is code that forms part of my quicklook software. In some ways I am
submitting it here as a test for the repo, however I am as yet undecided
how I will separate the different sections of this software between the
different instruments and I may well set this up so that there are
chunks of code separated by instrument as libraries or modules or
plugins or something. So the presence of this code here may well be
permanent.

The code basically provides the ability to read in ceilometer messages.
Currently it can only read Message2 formatted messages. The code can be
used to read the header first, then use this to check the header type.
Asuming it is a Message2, the message can be read and the data
extracted.